core: add missing override keyword to related classes
Reviewed-on: https://gerrit.st.com/c/linuxandroidopen/stm-hal-iio/+/257170
Tested-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>

+++ core/HWSensorBase.cpp
@@ -827,7 +827,7 @@ HWSensorBaseWithPollrate::HWSensorBaseWithPollrate(HWSensorBaseCommonData *data,
                                                    float power_consumption, int module)
     : HWSensorBase(data, name, handle, sensor_type, hw_fifo_len, power_consumption, module)
 {
-    float min_sampling_frequency = UINT_MAX, max_sampling_frequency = 0;
+    float min_sampling_frequency = FLT_MAX, max_sampling_frequency = 0;
     unsigned int i;
 
     memcpy(&sampling_frequency_available, sfa, sizeof(sampling_frequency_available));
